refactor: cutout Prefix to single file
https://github.com/akky-xxxx/eslint-plugin-strict-check/issues/285
