+++ src/shared/const/Prefix/index.ts
@@ -1,0 +1,1 @@
+export const Prefix = "strict-check"
