+++ src/shared/const/index.ts
@@ -1,4 +1,3 @@
 export const Off = 0
 export const Warn = 1
 export const Error = 2
-export const Prefix = "strict-check"
